Full Image Processing Codebase
This has handling code/scheduling code for the arduino, and cude parallelized processing for a jetson. It uses the PSF, deconvolution, and Weiner denoising. It produces a low res image first, then increasingly adds detail in high contrast sections of the image as it continues. It also handles faults, power management, and virtualized data redundancy similar to a RAID config, but with a single drive using separate partitions. Lastly, it includes placeholders for systems that aren't fully fleshed out yet.